change from long to int for processId

+++ model/src/main/java/com/ftn/sbnz/model/events/IOEvent.java
@@ -11,5 +11,5 @@ import org.kie.api.definition.type.Role;
 @AllArgsConstructor
 public class IOEvent {
 
-    private long processId;
+    private int processId;
 }
